Add calculation feature
I have succefully added the calculation feature I had in mind. TOP
provides a student example that has the same functionality as mine, but
it appears to be a little more simple. It won't allow you to fully input
a long mathematical expression that would be incorrect.

I wrote an example into the comments of my code that would produce an
incorrect value because it evaluates the code from left to right without
respect to traditional order of operations. After thinking about this,
I'm going to open a new branch in my code to rectify that and make it so
that only single operations are shown at once.